Separerer koin profilene i egne filer
